Improvements to the AclQueryAugmentor.
Introduced WhereClause value object to easily build where clauses and capture parameters. We now also guard the execution of an update for previously unsaved objects by checking whether we we find at least one for-update permission on the given domain type.

Improved test for rejected save on new object by checking the number of items in the store are the same before and after the attempt to save the object. Also checking that no identifier has been assigned to the instance.